♻️ Refactor debugger to improve testability
